+++ src/module/rules/rule-element/base-speed.ts
@@ -8,19 +8,6 @@ import { ModelPropsFromRESchema, ResolvableValueField, RuleElementSchema, RuleEl
 import fields = foundry.data.fields;
 
 const SPEED_TYPE_PATTERN = /movement\.speeds\.(land|burrow|climb|fly|swim)/g;
-const SPEED_VALUE_PATTERN = /movement\.speeds\.(land|burrow|climb|fly|swim)\.value/;
-
-/**
- * How a BaseSpeed formula that reads another speed's total is built in prepareMovementData.
- *
- * - equal: Resolved number matches the parent speed's total (e.g. fly = land.value). Treat as the same
- *   speed: reuse the parent's base and modifier domains so shared bonuses apply once and stack normally.
- * - scaled: Formula uses parent.value but the result differs (e.g. land.value * 0.5). Parent modifiers are
- *   already in that number; only apply modifiers on this movement type's domain afterward.
- * - independent: Fixed value, reads .base only, or min/max chose a constant floor/cap. Build like a normal
- *   speed with full modifier domains (all-speeds, speed, type-speed).
- */
-type SpeedDeriveKind = "equal" | "scaled" | "independent";
 
 function getSpeedFormulaDependsOn(formula: unknown, selfType: MovementType): MovementType[] {
     if (typeof formula !== "string") return [];
@@ -30,31 +17,6 @@ function getSpeedFormulaDependsOn(formula: unknown, selfType: MovementType): Mov
         if (tupleHasValue(MOVEMENT_TYPES, type) && type !== selfType) found.add(type);
     }
     return [...found];
-}
-
-function getDeriveParentType(formula: string | number, dependsOn: MovementType[]): MovementType | null {
-    if (dependsOn.length === 0) return null;
-    if (typeof formula === "string") {
-        const fromValue = dependsOn.find((type) => formula.includes(`speeds.${type}.value`));
-        if (fromValue) return fromValue;
-    }
-    return dependsOn[0] ?? null;
-}
-
-function classifySpeedDeriveKind(
-    formula: string | number,
-    resolved: number,
-    parentValue: number | undefined,
-): SpeedDeriveKind {
-    if (typeof formula !== "string" || parentValue === undefined || !SPEED_VALUE_PATTERN.test(formula)) {
-        return "independent";
-    }
-    if (resolved === parentValue) return "equal";
-    if (/\b(?:min|max)\s*\(/i.test(formula)) {
-        const literals = [...formula.matchAll(/(?<![.\w])(\d+)(?![.\w])/g)].map((m) => Number(m[1]));
-        if (literals.includes(resolved)) return "independent";
-    }
-    return "scaled";
 }
 
 /**
@@ -95,6 +57,7 @@ class BaseSpeedRuleElement extends RuleElement<BaseSpeedRuleSchema> {
         synthetics.push({
             dependsOn: getSpeedFormulaDependsOn(this.value, speedType),
             deferred: this.#createMovementType(speedType),
+            test: (options = {}) => this.test(options.test ?? []),
         });
     }
 
@@ -103,7 +66,7 @@ class BaseSpeedRuleElement extends RuleElement<BaseSpeedRuleSchema> {
         return (options: { test?: string[] | Set<string> } = {}): BaseSpeedSynthetic | null => {
             if (!this.test(options.test ?? [])) return null;
             const value = Math.trunc(Number(this.resolveValue(this.value)));
-            if (!(value > 0)) {
+            if (!(value > 0 || (this.force && value === 0))) {
                 if (!Number.isInteger(value)) this.failValidation("Failed to resolve value");
                 return null;
             }
@@ -129,4 +92,4 @@ type BaseSpeedRuleSchema = RuleElementSchema & {
     force: fields.BooleanField<boolean, boolean, false, false, true>;
 };
 
-export { BaseSpeedRuleElement, classifySpeedDeriveKind, getDeriveParentType };
+export { BaseSpeedRuleElement };
