+++ static/lang/en.json
@@ -669,7 +669,8 @@
                     "Fly": "Fly",
                     "Label": "{type} Speed",
                     "Land": "Land",
-                    "Swim": "Swim"
+                    "Swim": "Swim",
+                    "Travel": "Travel"
                 }
             },
             "NPC": {
